+++ app/src/main/java/nextstep/jwp/framework/manager/annotation/Controller.java
@@ -1,12 +1,11 @@
-package nextstep.jwp.manager.annotation;
+package nextstep.jwp.framework.manager.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Target(ElementType.METHOD)
+@Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface PostMapping {
-    String value();
+public @interface Controller {
 }
