+++ app/src/main/java/nextstep/jwp/framework/manager/annotation/GetMapping.java
@@ -1,11 +1,12 @@
-package nextstep.jwp.manager.annotation;
+package nextstep.jwp.framework.manager.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Target(ElementType.TYPE)
+@Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface Controller {
+public @interface GetMapping {
+    String value();
 }
