+++ app/src/main/java/nextstep/jwp/framework/manager/annotation/PostMapping.java
@@ -1,12 +1,12 @@
-package nextstep.jwp.manager.annotation;
+package nextstep.jwp.framework.manager.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Target(ElementType.PARAMETER)
+@Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface RequestParameter {
+public @interface PostMapping {
     String value();
 }
