+++ app/src/main/java/nextstep/jwp/framework/manager/annotation/RequestParameter.java
@@ -1,12 +1,12 @@
-package nextstep.jwp.manager.annotation;
+package nextstep.jwp.framework.manager.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Target(ElementType.METHOD)
+@Target(ElementType.PARAMETER)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface GetMapping {
+public @interface RequestParameter {
     String value();
 }
