+++ app/src/main/java/nextstep/jwp/framework/request/HttpMethod.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.request;
+package nextstep.jwp.framework.request;
 
 import java.util.Arrays;
