+++ app/src/main/java/nextstep/jwp/framework/response/HttpStatusCode.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.response;
+package nextstep.jwp.framework.response;
 
 public enum HttpStatusCode {
