+++ app/src/main/java/nextstep/jwp/framework/response/ServerResponse.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.response;
+package nextstep.jwp.framework.response;
 
 import org.apache.commons.io.FilenameUtils;
