+++ app/src/test/java/nextstep/jwp/framework/request/HttpMethodTest.java
@@ -1,11 +1,10 @@
-package nextstep.jwp.request;
+package nextstep.jwp.framework.request;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.*;
 
 class HttpMethodTest {
